# core/generated/swagger-ui/schema-registry-api-spec.yaml
openapi: 3.0.1
info:
  title: Confluent Schema Registry
  version: v1
paths:
...
  /subjects:
    get:
      summary: List subjects
      description: Retrieves a list of registered subjects matching specified parameters.
      operationId: list
      parameters:
      - name: subjectPrefix
        in: query
        description: Subject name prefix
        schema:
          type: string
          default: ':*:'
      - name: deleted
        in: query
        description: Whether to look up deleted subjects
        schema:
          type: boolean
      - name: deletedOnly
        in: query
        description: Whether to return deleted subjects only
        schema:
          type: boolean
      responses:
        "200":
          description: The subjects matching the specified parameters
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: string
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: Error code 50001 -- Error in the backend datastore
  /subjects/{subject}:
    post:
      summary: Lookup schema under subject
      description: "Check if a schema has already been registered under the specified\
        \ subject. If so, this returns the schema string along with its globally unique\
        \ identifier, its version under this subject and the subject name."
      operationId: lookUpSchemaUnderSubject
      parameters:
      - name: subject
        in: path
        description: Subject under which the schema will be registered
        required: true
        schema:
          type: string
      - name: normalize
        in: query
        description: Whether to lookup the normalized schema
        schema:
          type: boolean
      - name: deleted
        in: query
        description: Whether to lookup deleted schemas
        schema:
          type: boolean
      requestBody:
        description: Schema
        content:
          application/vnd.schemaregistry.v1+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/vnd.schemaregistry+json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
          application/octet-stream:
            schema:
              $ref: '#/components/schemas/RegisterSchemaRequest'
        required: true
      responses:
        "200":
          description: The schema
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/Schema'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/Schema'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/Schema'
        "404":
          description: |-
            Error code 40401 -- Subject not found
            Error code 40403 -- Schema not found
        "500":
          description: Internal server error
    delete:
      summary: Delete subject
      description: Deletes the specified subject and its associated compatibility
        level if registered. It is recommended to use this API only when a topic needs
        to be recycled or in development environment.
      operationId: deleteSubject
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: permanent
        in: query
        description: Whether to perform a permanent delete
        schema:
          type: boolean
      responses:
        "200":
          description: Operation succeeded. Returns list of schema versions deleted
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
        "404":
          description: Error code 40401 -- Subject not found
        "500":
          description: Error code 50001 -- Error in the backend datastore
...
components:
  schemas:
...
    RegisterSchemaRequest:
      type: object
      properties:
        version:
          type: integer
          description: Version number
          format: int32
        id:
          type: integer
          description: Globally unique identifier of the schema
          format: int32
        schemaType:
          type: string
          description: Schema type
        references:
          type: array
          description: References to other schemas
          items:
            $ref: '#/components/schemas/SchemaReference'
        schema:
          type: string
          description: Schema definition string
      description: Schema register request
    SchemaReference:
      type: object
      properties:
        name:
          type: string
          description: Reference name
        subject:
          type: string
          description: Name of the referenced subject
        version:
          type: integer
          description: Version number of the referenced subject
          format: int32
      description: Schema reference
...
    Schema:
      type: object
      properties:
        subject:
          type: string
          description: Name of the subject
        version:
          type: integer
          description: Version number
          format: int32
        id:
          type: integer
          description: Globally unique identifier of the schema
          format: int32
        schemaType:
          type: string
          description: Schema type
        references:
          type: array
          description: References to other schemas
          items:
            $ref: '#/components/schemas/SchemaReference'
        schema:
          type: string
          description: Schema definition string
      description: Schema
...